cxx-qt-build: move write of exported include directories to interface
